Ha master

+++ backend/app_purchase/views.py
@@ -35,7 +35,6 @@ def get_quotation_list(request):
 				'supplier_name':sm.supplier.name,
 			}
 			result.append(res)
-
 		return JsonResponse({'msg': 200, 'result': result})
 	else:
 		return JsonResponse({'msg': 'Please use POST', 'result': 'null'})
@@ -48,34 +47,38 @@ def add_quotation(request):
 	:param request:(material_id, supplier_id, price)
 	:return: 200 or error message
 	"""
-	if request.method == 'POST':
-		params = request.POST
-		# 得到所有参数
-		material_id = params.get('material_id')
-		supplier_id = params.get('supplier_id')
-		price = params.get('price')
-		# 只有所有参数都收到且不为空
-		if material_id and supplier_id and price:
-			try:
-				sp = Supplier.objects.get(id=supplier_id)
-				mt = Material.objects.get(id=material_id)
-			except Supplier.DoesNotExist:
-				return JsonResponse({'msg': 'there is doesn`t have this supplier'})
-			except Material.DoesNotExist:
-				# 捕获找不到对象异常
-				return JsonResponse({'msg': 'there is doesn`t have this material'})
+	user=request.user
+	if user.has_perm('add_SupplierMaterial'):
+		if request.method == 'POST':
+			params = request.POST
+			# 得到所有参数
+			material_id = params.get('material_id')
+			supplier_id = params.get('supplier_id')
+			price = params.get('price')
+			# 只有所有参数都收到且不为空
+			if material_id and supplier_id and price:
+				try:
+					sp = Supplier.objects.get(id=supplier_id)
+					mt = Material.objects.get(id=material_id)
+				except Supplier.DoesNotExist:
+					return JsonResponse({'msg': 'there is doesn`t have this supplier'})
+				except Material.DoesNotExist:
+					# 捕获找不到对象异常
+					return JsonResponse({'msg': 'there is doesn`t have this material'})
 
-			sm = SupplierMaterial(price=price, material=mt, supplier=sp)
-			try:
-				sm.save()
-				return JsonResponse({'msg': 200, 'result': 'success'})
-			except IntegrityError:
-				# 捕获主键唯一冲突异常
-				return JsonResponse({'msg': 'this supplier has already bid'})
+				sm = SupplierMaterial(price=price, material=mt, supplier=sp)
+				try:
+					sm.save()
+					return JsonResponse({'msg': 200, 'result': 'success'})
+				except IntegrityError:
+					# 捕获主键唯一冲突异常
+					return JsonResponse({'msg': 'this supplier has already bid'})
+			else:
+				return JsonResponse({'msg': 'Incomplete parameters'})
 		else:
-			return JsonResponse({'msg': 'Incomplete parameters'})
+			return JsonResponse({'msg': 'Please use POST'})
 	else:
-		return JsonResponse({'msg': 'Please use POST'})
+		return JsonResponse({'msg': 'Permission Denied'})
 
 
 @method_decorator(csrf_exempt)
@@ -83,7 +86,7 @@ def get_quotation_detail(request):
 	"""
 	获取供应商报价详细信息
 	:param request: quotation_id
-	:return: 
+	:return:
 	"""
 	if request.method == 'POST':
 		params = request.POST
@@ -118,24 +121,28 @@ def delete_quotation(request):
 	:param request: quotation_id
 	:return: 200 or error message
 	"""
-	if request.method == 'POST':
-		params = request.POST
-		# 得到所有参数
-		quotation_id = params.get('quotation_id')
-		# 判断参数是否存在
-		if quotation_id:
-			try:
-				sm = SupplierMaterial.objects.get(id=quotation_id)
-			except SupplierMaterial.DoesNotExist:
-				# 捕获get不到的异常
-				return JsonResponse({'msg': 'this quotation does not exist'})
-			# normal
-			sm.delete()
-			return JsonResponse({'msg': 200, 'result': 'success'})
+	user=request.user
+	if user.has_perm('delete_SupplierMaterial'):
+		if request.method == 'POST':
+			params = request.POST
+			# 得到所有参数
+			quotation_id = params.get('quotation_id')
+			# 判断参数是否存在
+			if quotation_id:
+				try:
+					sm = SupplierMaterial.objects.get(id=quotation_id)
+				except SupplierMaterial.DoesNotExist:
+					# 捕获get不到的异常
+					return JsonResponse({'msg': 'this quotation does not exist'})
+				# normal
+				sm.delete()
+				return JsonResponse({'msg': 200, 'result': 'success'})
+			else:
+				return JsonResponse({'msg': 'Incomplete parameters'})
 		else:
-			return JsonResponse({'msg': 'Incomplete parameters'})
+			return JsonResponse({'msg': 'Please use POST'})
 	else:
-		return JsonResponse({'msg': 'Please use POST'})
+		return JsonResponse({'msg': 'Permission Denied'})
 
 
 @method_decorator(csrf_exempt)
@@ -154,7 +161,7 @@ def quotation_query(request):
 		where_args = {}
 		# 都有什么条件
 		if material_id:
-			where_args['material__id'] = material_id
+    			where_args['material__id'] = material_id
 		if supplier_name:
 			where_args['supplier__name'] = supplier_name
 		if material_name:
@@ -182,24 +189,33 @@ def update_quotation_price(request):
 	"""
 	更新物料报价信息, 传入报价编号和准备修改的名字
 	"""
-	if request.method == 'POST':
-		params = request.POST
-		# 得到所有参数
-		quotation_id = params.get('quotation_id')
-		price = params.get('new_price')
-		# 判断是否参数完整
-		if quotation_id and price:
-			try:
-				sm = SupplierMaterial.objects.get(id=quotation_id)
-			except SupplierMaterial.DoesNotExist:
-				return JsonResponse({'msg': 'this quotation does not exist'})
-			sm.price = price
-			sm.save()
-			return JsonResponse({'msg': 200, 'result': 'ok'})
+	user=request.user
+	if user.has_perm('modify_SupplierMaterial'):
+		if request.method == 'POST':
+			params = request.POST
+			# 得到所有参数
+			quotation_id = params.get('quotation_id')
+			price = params.get('new_price')
+			# 判断是否参数完整
+			if quotation_id and price:
+				try:
+					sm = SupplierMaterial.objects.get(id=quotation_id)
+				except SupplierMaterial.DoesNotExist:
+					return JsonResponse({'msg': 'this quotation does not exist'})
+				price = float(price)
+				sm.price = price
+				try:
+					sm.save()
+					return JsonResponse({'msg': 200, 'result': 'ok'})
+				except Exception as E:
+					print(E)
+					return JsonResponse({'msg':500, 'result':'未知错误'})
+			else:
+				return JsonResponse({'msg': 'update failed, Incomplete parameters'})
 		else:
-			return JsonResponse({'msg': 'update failed, Incomplete parameters'})
+			return JsonResponse({'msg': 'Please use POST', 'result': 'null'})
 	else:
-		return JsonResponse({'msg': 'Please use POST', 'result': 'null'})
+		return JsonResponse({'msg': 'Permission Denied'})
 
 
 @method_decorator(csrf_exempt)
@@ -207,20 +223,25 @@ def add_supplier(request):
 	"""
 	新增提供商
 	"""
-	if request.method == 'POST':
-		params = request.POST
-		# 得到所有参数
-		supplier_name = params.get('supplier_name')
-		supplier_phone = params.get('supplier_phone')
-		supplier_address = params.get('supplier_address')
-		if supplier_address and supplier_name and supplier_phone:
-			sp = Supplier(name=supplier_name, contact='no', phonenumber=supplier_phone, address=supplier_address)
-			sp.save()
-			return JsonResponse({'msg': 200})
+	user=request.user
+	if user.has_perm('add_Supplier'):
+		if request.method == 'POST':
+			params = request.POST
+			# 得到所有参数
+			supplier_name = params.get('supplier_name')
+			supplier_phone = params.get('supplier_phone')
+			supplier_address = params.get('supplier_address')
+			print(supplier_name)
+			if supplier_address and supplier_name and supplier_phone:
+				sp = Supplier(name=supplier_name, contact='no', phonenumber=supplier_phone, address=supplier_address)
+				sp.save()
+				return JsonResponse({'msg': 200})
+			else:
+				return JsonResponse({'msg': 'Incomplete parameters'})
 		else:
-			return JsonResponse({'msg': 'Incomplete parameters'})
+			return JsonResponse({'msg': 'Please use POST', 'result': 'null'})
 	else:
-		return JsonResponse({'msg': 'Please use POST', 'result': 'null'})
+		return JsonResponse({'msg': 'Permission Denied'})
 
 
 def get_lack_list(request):
@@ -304,93 +325,99 @@ def add_purchase(request):
 	:param: 采购人,审核人,采购商品列表字典items(matreial_id,num,supplier_id)
 	:return: 200 or error message
 	"""
-	if request.method == 'POST':
-		params = request.POST
-		print(params)
-		purchaser = params.get('purchaser')
-		checker = params.get('checker')
-		supplier_id = params.get('supplier')
-		material_id = params.get('material_id')
-		num = params.get('number')
-		if purchaser and checker and supplier_id and material_id and num:
-			# insert into db
-			price = get_min_price(material_id)
-			print(price)
-			supplier = Supplier.objects.get(id=supplier_id)
-			purchase = Purchase(purchaser=purchaser, checker=checker, supplier=supplier, totalprice=price)
-			try:
-				purchase.save()
-			except IntegrityError:
-				return JsonResponse({'msg': ' primary key error'})
-
-			
-			try:
-				material = Material.objects.get(id=material_id)
-				purchase_item = PurchaseProduct(purchase=purchase, material=material, number=num, price=price)
-				purchase_item.save()
-				return JsonResponse({'msg': 200, 'result': 'success'})
-			except IntegrityError:		# 捕获主键唯一冲突异常
-				return JsonResponse({'msg': ' primary key error'})
-			except Material.DoesNotExist:		# 捕获get异常
-				return JsonResponse({'msg': 'this Material does not exist'})
-			except Supplier.DoesNotExist:
-				return JsonResponse({'msg': 'this Supplier does not exist'})
-
-		else:
-			return JsonResponse({'msg':'parameter error'})
-	else:
-		return JsonResponse({'msg': 'Please use POST', 'result': 'null'})
-	"""
-	if request.method == 'POST':
-		params = request.POST
-		purchaser = params.get('purchaser')
-		checker = params.get('checker')
-		supplier_id = params.get('supplier')
-		items = params.get('items')
-		if purchaser and checker and items and supplier_id:
-			# 计算总价
-			total_price = 0
-			# 先转成列表，再判断字典
-			items = list(items)
-			for item in items:
-				item_dict = dict(item)
-				material_id = int(item_dict['material_id'])
+	user=request.user
+	if user.has_perm('add_Purchase'):
+		if request.method == 'POST':
+			params = request.POST
+			print(params)
+			purchaser = params.get('purchaser')
+			checker = params.get('checker')
+			supplier_id = params.get('supplier')
+			material_id = params.get('material_id')
+			num = params.get('number')
+			if purchaser and checker and supplier_id and material_id and num:
+				# insert into db
 				price = get_min_price(material_id)
-				total_price += price
-			# insert into db
-			purchase = Purchase(purchaser=purchaser, checker=checker, supplier=supplier_id, totalprice=total_price)
-			try:
-				purchase.save()
-				print(1231231);
-			except IntegrityError:
-				return JsonResponse({'msg': ' primary key error'})
+				print(price)
+				if not price:
+	    				return JsonResponse({'msg': '该商品暂未报价'})
+				supplier = Supplier.objects.get(id=supplier_id)
+				purchase = Purchase(purchaser=purchaser, checker=checker, supplier=supplier, totalprice=price)
+				try:
+					purchase.save()
+				except IntegrityError:
+					return JsonResponse({'msg': ' primary key error'})
 
-			for item in items:
-				item_dict = dict(item)
-				matreial_id = item_dict['material_id']
-				num = item_dict['num']
-				price = get_min_price(material_id)
+				
+				try:
+					material = Material.objects.get(id=material_id)
+					purchase_item = PurchaseProduct(purchase=purchase, material=material, number=num, price=price)
+					purchase_item.save()
+					return JsonResponse({'msg': 200, 'result': 'success'})
+				except IntegrityError:		# 捕获主键唯一冲突异常
+					return JsonResponse({'msg': ' primary key error'})
+				except Material.DoesNotExist:		# 捕获get异常
+					return JsonResponse({'msg': 'this Material does not exist'})
+				except Supplier.DoesNotExist:
+					return JsonResponse({'msg': 'this Supplier does not exist'})
 
-				if supplier_id and matreial_id and num and price:
-					try:
-						matreial = Material.objects.get(id=matreial_id)
-						purchase_item = PurchaseProduct(purchase=purchase, matreial=matreial, num=num, price=price)
-						purchase_item.save()
-						return JsonResponse({'msg': 200, 'result': 'success'})
-					except IntegrityError:		# 捕获主键唯一冲突异常
-						return JsonResponse({'msg': ' primary key error'})
-					except Material.DoesNotExist:		# 捕获get异常
-						return JsonResponse({'msg': 'this Material does not exist'})
-					except Supplier.DoesNotExist:
-						return JsonResponse({'msg': 'this Supplier does not exist'})
-
-				else:
-					return JsonResponse({'msg': 'Incomplete parameters'})
+			else:
+				return JsonResponse({'msg':'parameter error'})
 		else:
-			return JsonResponse({'msg':'parameter error'})
+			return JsonResponse({'msg': 'Please use POST', 'result': 'null'})
+		"""
+		if request.method == 'POST':
+			params = request.POST
+			purchaser = params.get('purchaser')
+			checker = params.get('checker')
+			supplier_id = params.get('supplier')
+			items = params.get('items')
+			if purchaser and checker and items and supplier_id:
+				# 计算总价
+				total_price = 0
+				# 先转成列表，再判断字典
+				items = list(items)
+				for item in items:
+					item_dict = dict(item)
+					material_id = int(item_dict['material_id'])
+					price = get_min_price(material_id)
+					total_price += price
+				# insert into db
+				purchase = Purchase(purchaser=purchaser, checker=checker, supplier=supplier_id, totalprice=total_price)
+				try:
+					purchase.save()
+					print(1231231);
+				except IntegrityError:
+					return JsonResponse({'msg': ' primary key error'})
+
+				for item in items:
+					item_dict = dict(item)
+					matreial_id = item_dict['material_id']
+					num = item_dict['num']
+					price = get_min_price(material_id)
+
+					if supplier_id and matreial_id and num and price:
+						try:
+							matreial = Material.objects.get(id=matreial_id)
+							purchase_item = PurchaseProduct(purchase=purchase, matreial=matreial, num=num, price=price)
+							purchase_item.save()
+							return JsonResponse({'msg': 200, 'result': 'success'})
+						except IntegrityError:		# 捕获主键唯一冲突异常
+							return JsonResponse({'msg': ' primary key error'})
+						except Material.DoesNotExist:		# 捕获get异常
+							return JsonResponse({'msg': 'this Material does not exist'})
+						except Supplier.DoesNotExist:
+							return JsonResponse({'msg': 'this Supplier does not exist'})
+
+					else:
+						return JsonResponse({'msg': 'Incomplete parameters'})
+			else:
+	    			return JsonResponse({'msg':'parameter error'})
+		else:
+			return JsonResponse({'msg': 'Please use POST', 'result': 'null'})
+		"""
 	else:
-		return JsonResponse({'msg': 'Please use POST', 'result': 'null'})
-	"""
+		return JsonResponse({'msg': 'Permission Denied'})
 
 
 @method_decorator(csrf_exempt)
@@ -400,19 +427,23 @@ def delete_purchases(request):
 	:param request:(id1,id2,id3)以逗号分割的字符串
 	:return: 200 or error message
 	"""
-	if request.method == 'POST':
-		ids = request.POST.get('ids')
-		if ids:
-			ids = ids.split(',')
+	user=request.user
+	if user.has_perm('delete_Purchase'):
+		if request.method == 'POST':
+			ids = request.POST.get('ids')
+			if ids:
+				ids = ids.split(',')
 
-		q = Q()
-		q.connector = 'OR'
-		for id in ids:
-			q.children.append(('id', id))
-		Purchase.objects.filter(q).delete()
-		return JsonResponse({'msg': 200, 'result': 'ok'})
+			q = Q()
+			q.connector = 'OR'
+			for id in ids:
+				q.children.append(('id', id))
+			Purchase.objects.filter(q).delete()
+			return JsonResponse({'msg': 200, 'result': 'ok'})
+		else:
+			return JsonResponse({'msg': 'Please use POST', 'result': 'null'})
 	else:
-		return JsonResponse({'msg': 'Please use POST', 'result': 'null'})
+		return JsonResponse({'msg': 'Permission Denied'})
 
 
 @method_decorator(csrf_exempt)
